requirements.txt
Django==2.2.5
djangorestframework==3.10.3
pytest==5.1.3